updating code style
- major refactor for mvvm group 
- minor refactor for viper
