--- Templates/File Template/MVVM/MVVM with Repositories.xctemplate/___FILEBASENAME___ViewModelUseCases.swift
@@ -1,15 +1,0 @@
-//
-//  ___FILENAME___
-//  ___PROJECTNAME___
-//
-//  Created by ___FULLUSERNAME___ on ___DATE___.
-//  ___COPYRIGHT___
-//
-
-import RxCocoa
-import RxSwift
-
-// MARK: - ___VARIABLE_sceneName___ViewModelImpl + RepositoryUseCases
-extension ___VARIABLE_sceneName___ViewModelImpl {
-    typealias RepositoryUseCases = 
-}
