+++ Templates/File Template/MVVM/MVVM+Routing.xctemplate/TemplateInfo.plist
@@ -3,9 +3,9 @@
 <plist version="1.0">
 <dict>
 	<key>DefaultCompletionName</key>
-	<string>MVVM + Repository</string>
+	<string>MVVM+Routing</string>
 	<key>Description</key>
-	<string>Contains base module for files using MVVM and Repository Patterns</string>
+	<string>Contains base module for files using MVVM with Routing Pattern</string>
 	<key>Kind</key>
 	<string>Xcode.IDEKit.TextSubstitutionFileTemplateKind</string>
 	<key>Options</key>
@@ -62,20 +62,6 @@
 		</dict>
 		<dict>
 			<key>Default</key>
-			<string>___VARIABLE_sceneName:identifier___Repository</string>
-			<key>Description</key>
-			<string>The view Repository name</string>
-			<key>Identifier</key>
-			<string>repositoryName</string>
-			<key>Name</key>
-			<string>View Repository Name:</string>
-			<key>Required</key>
-			<true/>
-			<key>Type</key>
-			<string>static</string>
-		</dict>
-		<dict>
-			<key>Default</key>
 			<string>___VARIABLE_sceneName:identifier___ViewController</string>
 			<key>Description</key>
 			<string>The view controller name</string>
@@ -97,6 +83,20 @@
 			<string>viewName</string>
 			<key>Name</key>
 			<string>Storyboard view name:</string>
+			<key>Required</key>
+			<true/>
+			<key>Type</key>
+			<string>static</string>
+		</dict>
+		<dict>
+			<key>Default</key>
+			<string>___VARIABLE_sceneName:identifier___Route</string>
+			<key>Description</key>
+			<string>View route name</string>
+			<key>Identifier</key>
+			<string>routerName</string>
+			<key>Name</key>
+			<string>View Route name:</string>
 			<key>Required</key>
 			<true/>
 			<key>Type</key>
