+++ Templates/File Template/MVVM/MVVM+Routing.xctemplate/___FILEBASENAME___Router.swift
