+++ Templates/File Template/MVVM/MVVM+Routing.xctemplate/___FILEBASENAME___ViewController.swift
@@ -1,0 +1,35 @@
+//
+//  ___FILENAME___
+//  ___PROJECTNAME___
+//
+//  Created by ___FULLUSERNAME___ on ___DATE___.
+//  ___COPYRIGHT___
+//
+
+import UIKit
+
+// MARK: - ___VARIABLE_sceneName___ViewController
+final class ___VARIABLE_sceneName___ViewController: UIViewController, StoryboardInitializable {
+
+    // - Outlets
+    
+    // - Internal Propreties
+    var viewModel: ___VARIABLE_sceneName___ViewModelType!
+
+    // - Private Properties
+    private let disposeBag = DisposeBag()
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        self.configUI()
+    }
+
+}
+
+// MARK: - ___VARIABLE_sceneName___ViewController + Private
+private extension ___VARIABLE_sceneName___ViewController {
+
+    func configUI() {
+
+    }
+}
