+++ Templates/File Template/MVVM/MVVM+Routing.xctemplate/___FILEBASENAME___ViewModel.swift
@@ -1,0 +1,24 @@
+//
+//  ___FILENAME___
+//  ___PROJECTNAME___
+//
+//  Created by ___FULLUSERNAME___ on ___DATE___.
+//  ___COPYRIGHT___
+//
+
+import Foundation
+
+// MARK: - ___VARIABLE_sceneName___ViewModelType
+public protocol ___VARIABLE_sceneName___ViewModelType {}
+
+// MARK: -  ___VARIABLE_sceneName___ViewModel
+final class ___VARIABLE_sceneName___ViewModel: ___VARIABLE_sceneName___ViewModelType {
+    
+    // - Private Properties
+    private let router: ___VARIABLE_sceneName___Navigation
+    
+    // - Base
+    init(router: ___VARIABLE_sceneName___Navigation) {
+        self.router = router
+    }
+}
