+++ Templates/File Template/MVVM/MVVM+Storyboard.xctemplate/TemplateIcon.png
