+++ Templates/File Template/MVVM/MVVM+Storyboard.xctemplate/___FILEBASENAME___ViewController.swift
@@ -6,36 +6,27 @@
 //  ___COPYRIGHT___
 //
 
-import RxCocoa
-import RxSwift
+import UIKit
 
 // MARK: - ___VARIABLE_sceneName___ViewController
 final class ___VARIABLE_sceneName___ViewController: UIViewController, StoryboardInitializable {
 
     // - Outlets
-    
-    // - Internal Propreties
-    var viewModel: ___VARIABLE_sceneName___ViewModel!
 
-    // - Private Properties
-    private let disposeBag = DisposeBag()
-    
+    // - Internal Properties
+    public var viewModel: ___VARIABLE_sceneName___ViewModelType!
+
     override func viewDidLoad() {
         super.viewDidLoad()
         self.configUI()
-        self.configBinding()
     }
 
 }
 
-// MARK: - ___VARIABLE_sceneName___ViewController + Private
+// MARK: - Private
 private extension ___VARIABLE_sceneName___ViewController {
 
     func configUI() {
-
-    }
-
-    func configBinding() {
 
     }
 }
