+++ Templates/File Template/MVVM/MVVM+Storyboard.xctemplate/___FILEBASENAME___ViewModel.swift
@@ -1,0 +1,19 @@
+//
+//  ___FILENAME___
+//  ___PROJECTNAME___
+//
+//  Created by ___FULLUSERNAME___ on ___DATE___.
+//  ___COPYRIGHT___
+//
+
+import Foundation
+
+public protocol ___VARIABLE_sceneName___ViewModelType {}
+
+final class ___VARIABLE_sceneName___ViewModel: ___VARIABLE_sceneName___ViewModelType {
+
+    public init() {
+        // call custom object's init here
+    }
+
+}
