--- Templates/File Template/MVVM/MVVM.xctemplate/___FILEBASENAME___ViewController.swift
@@ -1,37 +1,0 @@
-//
-//  ___FILENAME___
-//  ___PROJECTNAME___
-//
-//  Created by ___FULLUSERNAME___ on ___DATE___.
-//  ___COPYRIGHT___
-//
-
-import RxCocoa
-import RxSwift
-
-class ___VARIABLE_sceneName___ViewController: UIViewController, StoryboardInitializable {
-
-    // - Outlets
-
-    // - Internal Properties
-    var viewModel: ___VARIABLE_sceneName___ViewModel!
-
-    // - Private Properties
-    private var disposeBag = DisposeBag()
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        self.configUI()
-        self.configBinding()
-    }
-
-    // - Setup
-    private func configUI() {
-
-    }
-
-    private func configBinding() {
-
-    }
-
-}
