--- Templates/File Template/MVVM/MVVM.xctemplate/___FILEBASENAME___ViewModel.swift
@@ -1,21 +1,0 @@
-//
-//  ___FILENAME___
-//  ___PROJECTNAME___
-//
-//  Created by ___FULLUSERNAME___ on ___DATE___.
-//  ___COPYRIGHT___
-//
-
-import RxSwift
-import RxCocoa
-
-class ___VARIABLE_sceneName___ViewModel {
-    
-    // - Private Properties
-    private var disposeBag = DisposeBag()
-
-    init() {
-        // call custom object's init here
-    }
-
-}
