+++ Templates/File Template/MVVM/Repository.xctemplate/___FILEBASENAME___Repository.swift
@@ -6,16 +6,15 @@
 //  ___COPYRIGHT___
 //
 
-import RxCocoa
-import RxSwift
+// MARK: - ___FILEBASENAME___RepositoryType
+public protocol ___FILEBASENAME___RepositoryType {
+
+}
 
 // MARK: - ___FILEBASENAME___Repository
-final class ___FILEBASENAME___Repository {
+final public class ___FILEBASENAME___Repository: ___FILEBASENAME___RepositoryType {
     
 //    private let database:
 //    private let network:
 
 }
-
-// MARK: - ___FILEBASENAME___Repository + UseCase
-extension ___FILEBASENAME___Repository {}
