+++ Templates/File Template/MVVM/UseCase.xctemplate/___FILEBASENAME___UseCase.swift
@@ -9,8 +9,8 @@
 import RxCocoa
 import RxSwift
 
-// MARK: - ___FILEBASENAME___UseCase
-/// Abstract Use Case for ___FILEBASENAME___ module
-protocol ___FILEBASENAME___UseCase {
-    
-}
+// MARK: - ___FILEBASENAME___UseCaseType
+/// Abstract layer for ___FILEBASENAME___UseCase
+public protocol ___FILEBASENAME___UseCaseType { }
+
+final public class ___FILEBASENAME___UseCase: ___FILEBASENAME___UseCaseType {}
